Latest Commit - New Year
Latest Commit - New Year 2019

+++ controllers/ProjectController.php
@@ -5,6 +5,7 @@ namespace app\controllers;
 use app\components\AccessRule;
 use app\models\Activity;
 use app\models\ActivityFiles;
+use app\models\Consistency;
 use app\models\LastPage;
 use app\models\Managers;
 use app\models\Procced;
@@ -26,6 +27,7 @@ use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 use yii\web\Response;
 use yii\web\UploadedFile;
+use yii\helpers\ArrayHelper;
 
 /**
  * ProjectController implements the CRUD actions for Project model.
@@ -85,7 +87,7 @@ class ProjectController extends Controller
     {
         $searchModel = new ProjectSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
-        $dataProvider->query->where('created_by = ' . Yii::$app->user->identity->id);
+        //$dataProvider->query->where('created_by = ' . Yii::$app->user->identity->id);
         return $this->render('user_index', [
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
@@ -110,10 +112,23 @@ class ProjectController extends Controller
                 $items = Yii::$app->request->post();
                 $temp_project_kpi_id = 0;
                 $temp_project_plan_id = 0;
+                $temp_consistency_id = 0;
                 $temp_lastpage_id = 0;
 
                 //var_dump($items['Project']['temp_project_kpi_id']);
                 //var_dump($items['Project']['temp_project_plan_id']);
+                //var_dump($items['Project']['temp_project_consistency']);
+
+                foreach ($items['Project']['temp_project_consistency'] as $key => $val) {
+                    $project_consistency = new Consistency();
+                    $project_consistency->cons_strategic_id = $val['cons_strategic_id'];
+                    $project_consistency->cons_goal_id = $val['cons_goal_id'];
+                    $project_consistency->cons_strategy_id = $val['cons_strategy_id'];
+                    $project_consistency->cons_indicator_id = $val['cons_indicator_id'];
+                    $project_consistency->project_act_key = $randomString;
+                    $project_consistency->save();
+                    $temp_consistency_id = $project_consistency->consistency_id;
+                }
 
                 foreach ($items['Project']['temp_project_kpi_id'] as $key => $val) {
                     $project_kpi = new ProjectKpi();
@@ -156,6 +171,7 @@ class ProjectController extends Controller
             $project_paomai = new ProjectPaomai();
             $project_paomai->project_quantity = $model->paomai_quantity;
             $project_paomai->project_quality = $model->paomai_quality;
+            $project_paomai->project_time = $model->paomai_time;
             $project_paomai->save();
 
             $project_laksana = new ProjectLaksana();
@@ -164,6 +180,7 @@ class ProjectController extends Controller
             $project_laksana->save();
 
             $model->projecti_paomai_id = $project_paomai->paomai_id;
+            $model->project_consistency_id = $temp_consistency_id;
             $model->created_by = Yii::$app->user->identity->id;
             $model->project_status = Project::PROJECT_RUNNING;
             $model->project_laksana_id = $project_laksana->laksana_id;
@@ -174,7 +191,6 @@ class ProjectController extends Controller
             $model->temp_project_kpi_id = \yii\helpers\Json::encode($model->temp_project_kpi_id);
             $model->temp_project_plan_id = \yii\helpers\Json::encode($model->temp_project_plan_id);
             $model->lastpage_main = \yii\helpers\Json::encode($model->lastpage_main);
-
             if ($model->validate()){
                 $model->project_key = $randomString;
                 $model->save();
@@ -201,28 +217,52 @@ class ProjectController extends Controller
 
     public function actionUpdate($id)
     {
+        $user = Yii::$app->user->identity->id;
         $model = $this->findModel($id);
-        $model->temp_type = $model->projectLaksana->projectType->type_id;
-        $model->temp_procced = $model->projectLaksana->procced->procced_id;
+        $creator = $model->created_by;
+        $master = new Project();  
 
-        $model->temp_project_kpi_id = ProjectKpi::find()->where(['kpi_project_key' => $model->project_key])->all();
-        $model->temp_project_plan_id = ProjectPlan::find()->where(['plan_project_key' => $model->project_key])->all();
-        $model->lastpage_main = LastPage::find()->where(['project_act_key' => $model->project_key])->all();
+        if($creator == $user){
+            $model->temp_type = $model->projectLaksana->projectType->type_id;
+            $model->temp_procced = $model->projectLaksana->procced->procced_id;
 
-        $model->paomai_quantity = ProjectPaomai::find()->where(['paomai_id' => $model->projecti_paomai_id])->one()->project_quantity;
-        $model->paomai_quality = ProjectPaomai::find()->where(['paomai_id' => $model->projecti_paomai_id])->one()->project_quality;
+            $model->temp_project_kpi_id = ProjectKpi::find()->where(['kpi_project_key' => $model->project_key])->all();
+            $model->temp_project_plan_id = ProjectPlan::find()->where(['plan_project_key' => $model->project_key])->all();
+            $model->temp_project_consistency = Consistency::find()->where(['project_act_key' => $model->project_key])->all();
+            $model->lastpage_main = LastPage::find()->where(['project_act_key' => $model->project_key])->all();
+
+            $model->paomai_quantity = ProjectPaomai::find()->where(['paomai_id' => $model->projecti_paomai_id])->one()->project_quantity;
+            $model->paomai_quality = ProjectPaomai::find()->where(['paomai_id' => $model->projecti_paomai_id])->one()->project_quality;
+            $model->paomai_time = ProjectPaomai::find()->where(['paomai_id' => $model->projecti_paomai_id])->one()->project_time;
+        }else{
+            Yii::$app->getSession()->setFlash('project_update_denied', [
+                'type' =>  Growl::TYPE_DANGER,
+                'duration' => 3000,
+                'icon' => 'fa fa-close',
+                'title' => $model->project_name,
+                'message' => 'ไม่สามารถแก้ไขโครงการของคนอืนได้',
+                'positonY' => 'bottom',
+                'positonX' => 'right'
+            ]);
+            return $this->redirect(['index']);
+        }
 
         if ($model->load(Yii::$app->request->post())) {
             $project_kpi = ProjectKpi::find()->where(['kpi_project_key' => $model->project_key])->all();
             $project_plan = ProjectPlan::find()->where(['plan_project_key' => $model->project_key])->all();
             $lastpages = LastPage::find()->where(['project_act_key' => $model->project_key])->all();
+            $consistencies = Consistency::find()->where(['project_act_key' => $model->project_key])->all();
             $transaction = Yii::$app->db->beginTransaction();
             try {
                 $items = Yii::$app->request->post();
                 $temp_project_kpi_id = 0;
                 $temp_project_plan_id = 0;
                 $temp_lastpage_id = 0;
+                $temp_consistency_id = 0;
 
+                foreach($consistencies as $delete){
+                    $delete->delete();
+                }
                 foreach($project_kpi as $delete){
                     $delete->delete();
                 }
@@ -231,6 +271,17 @@ class ProjectController extends Controller
                 }
                 foreach($lastpages as $delete3){
                     $delete3->delete();
+                }
+
+                foreach ($items['Project']['temp_project_consistency'] as $key => $val) {
+                    $project_consistency = new Consistency();
+                    $project_consistency->cons_strategic_id = $val['cons_strategic_id'];
+                    $project_consistency->cons_goal_id = $val['cons_goal_id'];
+                    $project_consistency->cons_strategy_id = $val['cons_strategy_id'];
+                    $project_consistency->cons_indicator_id = $val['cons_indicator_id'];
+                    $project_consistency->project_act_key = $model->project_key;
+                    $project_consistency->save();
+                    $temp_consistency_id = $project_consistency->consistency_id;
                 }
 
                 foreach ($items['Project']['temp_project_kpi_id'] as $key => $val) {
@@ -274,6 +325,7 @@ class ProjectController extends Controller
             $project_paomai = new ProjectPaomai();
             $project_paomai->project_quantity = $model->paomai_quantity;
             $project_paomai->project_quality = $model->paomai_quality;
+            $project_paomai->project_time = $model->paomai_time;
             $project_paomai->save();
 
             $project_laksana = new ProjectLaksana();
@@ -283,6 +335,7 @@ class ProjectController extends Controller
 
             $model->projecti_paomai_id = $project_paomai->paomai_id;
             $model->created_by = Yii::$app->user->identity->id;
+            $model->project_consistency_id = $temp_consistency_id;
             $model->project_status = Project::PROJECT_RUNNING;
             $model->project_laksana_id = $project_laksana->laksana_id;
             $model->project_kpi_id = $temp_project_kpi_id;
@@ -313,13 +366,30 @@ class ProjectController extends Controller
 
         return $this->render('update', [
             'model' => $model,
+            'tempGoal' => '',
+            'tempStrategy' => '',
+            'tempIndicator' =>  ''
         ]);
     }
 
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
-
+        $user = Yii::$app->user->identity->id;
+        $model = $this->findModel($id);
+        $creator = $model->created_by;
+        if($creator == $user){
+            $this->findModel($id)->delete();
+        }else{
+            Yii::$app->getSession()->setFlash('project_delete_denied', [
+                'type' =>  Growl::TYPE_DANGER,
+                'duration' => 3000,
+                'icon' => 'fa fa-close',
+                'title' => $model->project_name,
+                'message' => 'ไม่สามารถลบโครงการของคนอืนได้',
+                'positonY' => 'bottom',
+                'positonX' => 'right'
+            ]);
+        }
         return $this->redirect(['index']);
     }
 
@@ -338,6 +408,8 @@ class ProjectController extends Controller
         if (Yii::$app->request->isPost) {
             $model = Project::find()->where(['project_id' => $project_id])->one();
             $model->project_status = Project::PROJECT_FINISHED;
+            $model->temp_type = 0; // Temp
+            $model->temp_procced = 0; // Temp
             if($model->save()){
                 Yii::$app->getSession()->setFlash('project_accept_ok', [
                     'type' =>  Growl::TYPE_SUCCESS,
@@ -369,6 +441,8 @@ class ProjectController extends Controller
         if (Yii::$app->request->isPost) {
             $model = Project::find()->where(['project_id' => $project_id])->one();
             $model->project_status = Project::PROJECT_RUNNING;
+            $model->temp_type = 0; // Temp
+            $model->temp_procced = 0; // Temp
             if($model->save()){
                 Yii::$app->getSession()->setFlash('project_accept_ok', [
                     'type' =>  Growl::TYPE_SUCCESS,
@@ -420,14 +494,13 @@ class ProjectController extends Controller
     }
 
     public function actionPreview($project_id, $project_name) {
-        //$mpdf = new Mpdf(['mode' => 's']);
-        $mpdf = new \Mpdf\Mpdf(['tempDir' => __DIR__ . '/mpdf/temp']);
+        $mpdf = new Mpdf(['mode' => 's']);
         $model = $this->findModel($project_id);
          $content  = $this->renderPartial('project_print', [
             'model' => $model,
         ]);
 
-        //return $content;
+//        return $content;
 
         $stylesheet = "
         body{font-family: Garuda}
@@ -457,6 +530,53 @@ class ProjectController extends Controller
         $mpdf->WriteHTML($stylesheet,1);
         $mpdf->WriteHTML($content,2);
         $mpdf->Output($project_name . ' - ฉบับสมบูรณ์', 'D');
+    }
 
+    public function actionGetgoal()
+    {
+        $out = [];
+        $master = new Project();
+        if (isset($_POST['depdrop_parents'])) {
+            $parents = $_POST['depdrop_parents'];
+            if ($parents != null) {
+                $strategic_id = $parents[0];
+                $out = $master->getRelatedGoal($strategic_id);
+                echo Json::encode(['output' => $out, 'selected' => '']);
+                return;
+            }
+        }
+        echo Json::encode(['output' => '', 'selected' => '']);
+    }
+
+    public function actionGetstrategy()
+    {
+        $out = [];
+        $master = new Project();
+        if (isset($_POST['depdrop_parents'])) {
+            $parents = $_POST['depdrop_parents'];
+            if ($parents != null) {
+                $goal_id = $parents[0];
+                $out = $master->getRelatedStrategy($goal_id);
+                echo Json::encode(['output' => $out, 'selected' => '']);
+                return;
+            }
+        }
+        echo Json::encode(['output' => '', 'selected' => '']);
+    }
+    
+    public function actionGetindicator()
+    {
+        $out = [];
+        $master = new Project();
+        if (isset($_POST['depdrop_parents'])) {
+            $parents = $_POST['depdrop_parents'];
+            if ($parents != null) {
+                $goal_id = $parents[0];
+                $out = $master->getRelatedIndicator($goal_id);
+                echo Json::encode(['output' => $out, 'selected' => '']);
+                return;
+            }
+        }
+        echo Json::encode(['output' => '', 'selected' => '']);
     }
 }
